Factored out CDK agr_aws_env for reused accross components
