+++ pipeline/aws_infra/tests/unit/test_cdk_infra_stack.py
@@ -9,7 +9,14 @@ from aws_cdk.aws_config import ResourceType
 import aws_cdk.assertions as assertions
 
 from cdk_classes.cdk_infra_stack import CdkInfraStack
-from agr_aws_env import agr_aws_environment
+
+from pathlib import Path
+from sys import path as sys_path
+
+repo_root_path = Path(__file__).parent.parent.parent.parent
+sys_path.append(str(repo_root_path))
+
+from shared_aws_infra.agr_aws_env import agr_aws_environment  # noqa: E402
 
 app = App()
 stack = CdkInfraStack(app, "pytest-stack", env=agr_aws_environment)
